Converts commander commands to yargs commands
