+++ src/bin/commands/logout.js
@@ -1,0 +1,6 @@
+import {SETTINGS} from '../cli';
+
+exports.command = 'logout';
+exports.desc = 'Logout from Cryptid';
+exports.builder = {};
+exports.handler = () => SETTINGS.clear();
