+++ src/bin/commands/server.js
@@ -1,0 +1,6 @@
+import {LOGGER, SERVER} from '../cli';
+
+exports.command = 'server';
+exports.desc = 'Display currently configured server';
+exports.builder = {};
+exports.handler = () => LOGGER.info(SERVER);
