+++ src/bin/commands/user.js
@@ -1,0 +1,10 @@
+
+exports.command = 'user';
+exports.desc = 'User actions';
+exports.builder = (yargs) => {
+  yargs
+    .usage('Interact with user profile')
+    .commandDir('user')
+    .demandCommand();
+};
+exports.handler = () => {};
