--- src/bin/cryptid-logout.js
@@ -1,3 +1,0 @@
-import {SETTINGS} from './cli';
-
-SETTINGS.clear();
