--- src/bin/cryptid-properties.js
@@ -1,9 +1,0 @@
-import {loadHelp} from './help';
-import program from 'commander';
-
-program
-  .description(loadHelp('cryptid-properties'))
-  .command('list', 'List accessible properties')
-  .command('create', 'Create property')
-  .command('update', 'Update a property')
-  .parse(process.argv);
