--- src/bin/cryptid-server.js
@@ -1,5 +1,0 @@
-import {LOGGER, SERVER} from './cli';
-
-LOGGER.info(SERVER);
-
-process.exit(0);
