--- src/bin/cryptid-user.js
@@ -1,5 +1,0 @@
-import program from 'commander';
-
-program
-  .command('change-password', 'Change the password of the current user')
-  .parse(process.argv);
