--- src/bin/cryptid-version.js
@@ -1,4 +1,0 @@
-import {LOGGER} from './cli';
-import {loadVersion} from './help';
-
-LOGGER.info(loadVersion());
